refactor: webClientConf 제거 및 Util 변경
